made price field a decimalfield

+++ backend/backend/listings/models.py
@@ -7,7 +7,7 @@ FEEDBACK_CHOICES = [
 ]
 
 class Listing(models.Model):
-  price = models.IntegerField()
+  price = models.DecimalField(decimal_places=2, max_digits=9)
   title = models.CharField(max_length=25, unique=True)
   description = models.TextField(max_length=1000)
   created = models.DateTimeField(auto_now_add=True)
